+++ tests/test_connected_notebook_is_small.py
@@ -40,4 +40,6 @@ def test_offline_notebook_is_not_too_large(tmp_path):
     nb_py.write_text(text_notebook(connected=False))
     jupytext([str(nb_py), "--to", "ipynb", "--set-kernel", "itables", "--execute"])
     assert nb_ipynb.exists()
-    assert 825000 < nb_ipynb.stat().st_size < 875000
+    size_in_kb = nb_ipynb.stat().st_size // 1024
+
+    assert 850 < size_in_kb < 950