+++ docs/py/options/column_control.py
@@ -26,7 +26,7 @@ import itables
 
 itables.init_notebook_mode()
 
-df = itables.sample_dfs.get_countries(html=False)
+df = itables.sample_dfs.get_countries()
 
 # %% [markdown]
 # ## Getting started