+++ docs/py/options/scroll_y.py
@@ -24,5 +24,5 @@ import itables
 
 itables.init_notebook_mode()
 
-df = itables.sample_dfs.get_countries(html=False)
+df = itables.sample_dfs.get_countries()
 itables.show(df, scrollY="350px", scrollCollapse=True, paging=False)